requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
google-generativeai
